refactor: separate test code
