+++ tests/api/main.rs
@@ -1,0 +1,3 @@
+mod helpers;
+mod health_check;
+mod subscriptions;
